refactor(resources/bookmark): unified folder
